requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==4.2.20
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.13.2
whitenoise==6.9.0
